+++ src/clj/rems/util.clj
@@ -8,6 +8,9 @@
   (:import [clojure.lang Atom]
            [java.io ByteArrayOutputStream FileInputStream]))
 
+;; regex from https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/email#Validation
+(def +email-regex+ #"[a-zA-Z0-9.!#$%&'*+\/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*")
+
 (defn errorf
   "Throw a RuntimeException, args passed to `clojure.core/format`."
   [& fmt-args]